Refactor coherence policies and enhance event tracking in TP stages
- Updated coherence policies for collective stages (TPAllreduceStage, AllGatherStage, AllGatherVStage) to OUTPUT to ensure proper dirty marking and synchronization.
- Introduced a new EventRecord structure in MockBackend to track event operations (CREATE, DESTROY, RECORD, WAIT) for better test verification.
- Added methods to retrieve event records and counts for various operations in MockBackend.
- Created comprehensive unit tests for stream/coherence interactions, addressing two critical bugs related to coherence policies and stale events.
- Verified that mark_device_dirty_with_event() correctly replaces stale events, ensuring accurate synchronization during data transfers.
- Various correctness fixes on CUDA

+++ src/v2/utils/DebugEnv.h
@@ -2557,9 +2557,12 @@ namespace llaminar2
         bool tp_timing = false;      ///< Enable TP forward timing breakdown (env: LLAMINAR_TP_TIMING)
         bool skip_allreduce = false; ///< DIAGNOSTIC: Skip allreduce for profiling (env: LLAMINAR_SKIP_ALLREDUCE)
 
-        /// Allreduce precision: "fp16", "bf16", or "fp32" (env: LLAMINAR_ALLREDUCE_PRECISION, default: "fp16")
+        /// Global allreduce precision fallback: "fp16", "bf16", or "fp32"
+        /// (env: LLAMINAR_ALLREDUCE_PRECISION, default: "fp32")
+        /// This is the ULTIMATE FALLBACK — per-layer precision from the model schema
+        /// takes priority. Only used when no schema-level precision is configured.
         /// FP16/BF16 halves PCIe transfer bandwidth; FP32 is lossless but slower on bandwidth-limited links.
-        std::string allreduce_precision = "fp16";
+        std::string allreduce_precision = "fp32";
 
         /// Timeout in ms for TPWorkerPool::collectAll() (env: LLAMINAR_TP_COLLECT_TIMEOUT_MS, default: 0 = unlimited)
         /// 0 means wait forever (normal operation). Set to e.g. 30000 for a 30s safety net if debugging hangs.
